requirements.txt
opencv-python
numpy
matplotlib
tensorflow
scikit-learn
opencv-python
numpy
matplotlib
tensorflow
scikit-learn
streamlit
Pillow